Fixing "bitwise operation between different enumeration types" warnings.
git-svn-id: https://anonsvn.ncbi.nlm.nih.gov/repos/v1/trunk/c++@101580 78c7ea69-d796-4a43-9a09-de51944f1b03

+++ src/util/compress/api/archive.cpp
@@ -1131,7 +1131,7 @@ CArchiveFile::CArchiveFile(EFormat format, const string& filename)
 void CArchiveFile::Open(EAction action)
 {
     ARCHIVE_CHECK;
-    bool isread = (action & eRO) > 0;
+    bool isread = ((int)action & (int)eRO) > 0;
     if (isread) {
         ARCHIVE->OpenFile(m_FileName);
     } else {
@@ -1181,7 +1181,7 @@ void CArchiveMemory::Create(size_t initial_allocation_size)
 void CArchiveMemory::Open(EAction action)
 {
     ARCHIVE_CHECK;
-    bool isread = (action & eRO) > 0;
+    bool isread = ((int)action & (int)eRO) > 0;
     if (isread) {
         ARCHIVE->OpenMemory(m_Buf, m_BufSize);
     } else {
